requirements.txt
fastapi
uvicorn
pydantic
pyyaml
requests
# RAG (Phase 3 wiring later)
chromadb
# testing
pytest
